Revert to float32

+++ covid_forecasting_joint_learning/data/util.py
@@ -4,7 +4,7 @@ from . import cols as DataCol
 import torch
 
 
-DEFAULT_DTYPE = np.float64
+DEFAULT_DTYPE = np.float32
 
 
 def mkparent(path):
